largest prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -11,7 +11,7 @@
 
 int main(void)
 {
-	long long int num = 612852475143;
+	long int num = 612852475143;
 	int fac;
 	int largest_prime_fac = 1;
 	double square = sqrt(num);
